Fixed calculation

+++ ajan/model.py
@@ -14,7 +14,7 @@ class WorldModel(Model):
         for i in range(self.num_agents):
             x = random.randrange(self.grid.width)
             y = random.randrange(self.grid.height)
-            ethnicity = (((x+1) * (y+1)) / self.grid.width * self.grid.height) * len(Ethnicities)
+            ethnicity = ((x+1) * (y+1)) / float(self.grid.width * self.grid.height) * len(Ethnicities)
             a = PersonAgent(i, ethnicty=int(ethnicity))
             self.schedule.add(a)
             # Add the agent to a random grid cell
